리뷰 평균 별점 기능 완료, jsp 뿌려주기 완료

+++ pet-shopping-mall/src/main/java/com/shop/app/review/dto/ProductReviewAvgDto.java
@@ -1,11 +1,5 @@
 package com.shop.app.review.dto;
 
-import java.sql.Date;
-
-import com.shop.app.order.dto.OrderAdminListDto;
-import com.shop.app.product.entity.Product;
-import com.shop.app.review.entity.Review;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -19,5 +13,5 @@ public class ProductReviewAvgDto {
 	
 	private String productName; // product
 	private int productId; // product
-	private int reviewStarRate; // review
+	private double reviewStarRate; // review
 }
